feat: standardize voice dictation via shared hook and apply across pages\n\n- Add useVoiceDictation hook with robust MediaRecorder handling, MIME-type detection, iOS/Safari track cleanup, and error/debug logging\n- Refactor Playground SpeechToTextCard to use the shared hook (no behavior change, keeps debug + playback UI)\n- Refactor Issues NewTaskInput to use the same hook and append transcriptions to the description\n\nThis unifies voice dictation behavior and fixes recording issues on the main issues page by reusing the working Playground logic.
